Move cache type identifier helper functions
